requirements.txt
langchain-community
langchain-groq
requests
Flask
Flask-CORS

